garbageday_management/requirements.txt
asgiref==3.3.1
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-widget-tweaks==1.4.8
future==0.18.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
line-bot-sdk==1.18.0
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.6.0
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
reppy==0.4.14
requests==2.25.1
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
stripe==2.53.0
toml==0.10.2
tweepy==3.9.0
urllib3==1.26.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
pycrypto==2.6.1